Rename reset footprint to button
